requirements.txt
crewai
langchain_google_genai
load_dotenv
crewai_tools
crewai
langchain_groq
python-dotenv
